requirements.txt
asgiref==3.5.2
astroid==2.15.5
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
Django==4.1.2
django-adminlte-3==0.1.6
djangorestframework==3.15.1
executing==1.2.0
filelock==3.8.0
fonttools==4.39.4
ipykernel==6.23.1
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pylint==2.17.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2022.5
vboxapi==1.0
virtualenv==20.16.5
wcwidth==0.2.6
wrapt==1.15.0
